Сделал по PEP, разделил player от main
